+++ app/app.go
@@ -112,7 +112,7 @@ import (
 	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	mempool "github.com/cosmos/cosmos-sdk/types/mempool"
+	"github.com/cosmos/cosmos-sdk/types/mempool"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/cosmos/cosmos-sdk/types/msgservice"
 	sigtypes "github.com/cosmos/cosmos-sdk/types/tx/signing"
@@ -174,6 +174,8 @@ const (
 	FlagBlockedAddresses             = "blocked-addresses"
 	FlagUnsafeIgnoreBlockListFailure = "unsafe-ignore-block-list-failure"
 	FlagUnsafeDummyCheckTx           = "unsafe-dummy-check-tx"
+
+	FlagDisableOptimisticExecution = "cronos.disable-optimistic-execution"
 )
 
 var Forks = []Fork{}
@@ -410,7 +412,7 @@ func New(
 	})
 
 	blockSTMEnabled := cast.ToString(appOpts.Get(srvflags.EVMBlockExecutor)) == "block-stm"
-	optimisticExecutionDisabled := cast.ToString(appOpts.Get(srvflags.EVMOptimisticExecution)) == "disable"
+	optimisticExecutionDisabled := cast.ToBool(appOpts.Get(FlagDisableOptimisticExecution))
 
 	var cacheSize int
 	if !blockSTMEnabled && optimisticExecutionDisabled {