+++ integration_tests/test_mempool.py
@@ -9,6 +9,9 @@ from .utils import (
     CONTRACTS,
     KEYS,
     deploy_contract,
+    get_account_nonce,
+    replace_transaction,
+    send_transaction,
     send_txs,
     sign_transaction,
     wait_for_new_blocks,
@@ -138,3 +141,88 @@ def test_mempool_nonce(cronos_mempool):
             sender
         )
         assert orig_nonce + 4 + i == local_nonce
+
+
+@pytest.mark.flaky(max_runs=3)
+def test_tx_replacement(cronos_mempool):
+    w3 = cronos_mempool.w3
+    base_fee = w3.eth.get_block("latest")["baseFeePerGas"]
+    priority_fee = w3.eth.max_priority_fee
+    nonce = get_account_nonce(w3)
+    # replace with less than 10% bump, should fail
+    with pytest.raises(ValueError) as exc:
+        _ = replace_transaction(
+            w3,
+            {
+                "to": ADDRS["community"],
+                "value": 1,
+                "maxFeePerGas": base_fee + priority_fee,
+                "maxPriorityFeePerGas": priority_fee,
+                "nonce": nonce,
+                "from": ADDRS["validator"],
+            },
+            {
+                "to": ADDRS["community"],
+                "value": 2,
+                "maxFeePerGas": int((base_fee + priority_fee) * 1.05),  # +5% bump
+                "maxPriorityFeePerGas": int(priority_fee * 1.05),
+                "nonce": nonce,
+                "from": ADDRS["validator"],
+            },
+            KEYS["validator"],
+        )["transactionHash"]
+    assert "tx doesn't fit the replacement rule" in str(exc)
+
+    wait_for_new_blocks(cronos_mempool.cosmos_cli(), 1)
+    nonce = get_account_nonce(w3)
+    initial_balance = w3.eth.get_balance(ADDRS["community"])
+    # replace with more than 10% bump, should succeed
+    txhash = replace_transaction(
+        w3,
+        {
+            "to": ADDRS["community"],
+            "value": 3,
+            "maxFeePerGas": base_fee + priority_fee,
+            "maxPriorityFeePerGas": priority_fee,
+            "nonce": nonce,
+            "from": ADDRS["validator"],
+        },
+        {
+            "to": ADDRS["community"],
+            "value": 5,
+            "maxFeePerGas": int((base_fee + priority_fee) * 1.15),  # +15% bump
+            "maxPriorityFeePerGas": int(priority_fee * 1.15),
+            "nonce": nonce,
+            "from": ADDRS["validator"],
+        },
+        KEYS["validator"],
+    )["transactionHash"]
+    tx1 = w3.eth.get_transaction(txhash)
+    assert tx1["transactionIndex"] == 0
+    assert w3.eth.get_balance(ADDRS["community"]) == initial_balance + 5
+
+    # check that already accepted transaction cannot be replaced
+    txhash_noreplacemenet = send_transaction(
+        w3,
+        {
+            "to": ADDRS["community"],
+            "value": 10,
+            "maxFeePerGas": base_fee + priority_fee,
+            "maxPriorityFeePerGas": priority_fee,
+        },
+        KEYS["validator"],
+    )["transactionHash"]
+    tx2 = w3.eth.get_transaction(txhash_noreplacemenet)
+    assert tx2["transactionIndex"] == 0
+
+    with pytest.raises(ValueError) as exc:
+        w3.eth.replace_transaction(
+            txhash_noreplacemenet,
+            {
+                "to": ADDRS["community"],
+                "value": 15,
+                "maxFeePerGas": int((base_fee + priority_fee) * 1.15),  # +15% bump
+                "maxPriorityFeePerGas": int(priority_fee * 1.15),
+            },
+        )
+    assert "has already been mined" in str(exc)