+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ## v1.5.3
 
+### Bug fixes
 * [#1898](https://github.com/crypto-org-chain/cronos/pull/1898) Check authorisation list for blocklisted address.
 
 ## v1.5.2